Added map area selection extraction and summary info of the map area selection to the summary page.
Found out a way to find the selected latitude and longitude area on the map.
Summary page is now updated with this info. If there is no area selected there will be an error in that field, if there is an area selected it will display the lats and longs.
Multiple areas can still be selected, this needs to be fixed so only 1 rectangle can be selected.